Remove multi-user id from Stadia Store Links
Previously, the links to the Stadia Store contained "/u/1" in the links. This meant that when a user is logged in with multiple google accounts (e.g. private account + GSuite Work account), it would attempt to open the store page with the second (the one with index 1) account. 
This commit normalizes the linked store URLS and instead links to the store page without an account identifier.